requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.5
django-filter==2.4.0
djangorestframework==3.12.4
ecdsa==0.17.0
idna==2.10
install==1.3.4
Markdown==3.3.4
oauthlib==3.1.1
Pillow==8.3.1
pyasn1==0.4.8
pycparser==2.20
PyJWT==2.1.0
python-decouple==3.4
python-dotenv==0.18.0
python-jose==3.3.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
urllib3==1.26.6
